+++ CHANGELOG.md
@@ -6,3 +6,4 @@ I will record changes to this file just so i don't need to look at commit histor
 - Added kinto.py config file to repo.
 - Added homebrew to debian setup script.
 - Tweaked how vscode is installed on debian.
+- Added script for backing up and restoring config files.